@@ -5,45 +5,46 @@
   "main": "index.js",
   "scripts": {
     "test": "mocha",
+    "dev": "webpack-dev-server && nodemon server.js",
     "start": "node server.js"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "async": "^2.3.0",
     "axios": "^0.16.1",
     "babel-core": "^6.24.1",
     "babel-loader": "^7.0.0",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
     "body-parser": "^1.17.1",
-    "cheerio": "^0.22.0",
-    "cron": "^1.2.1",
     "express": "^4.15.2",
     "history": "^4.6.1",
     "mongoose": "^4.9.6",
     "morgan": "^1.8.1",
-    "nodemailer": "^4.0.1",
     "passport": "^0.3.2",
     "passport-http": "^0.3.0",
     "passport-local": "^1.0.0",
     "react": "^15.5.4",
     "react-dom": "^15.5.4",
+    "react-js-pagination": "^2.1.0",
+    "react-paginate": "^4.4.2",
     "react-redux": "^5.0.5",
     "react-router": "^4.1.1",
     "react-router-dom": "^4.1.1",
     "react-thunk": "^1.0.0",
     "redux": "^3.6.0",
     "redux-devtools-extension": "^2.13.2",
-    "redux-thunk": "^2.2.0",
-    "request": "^2.81.0",
-    "webpack": "^2.5.1"
+    "redux-thunk": "^2.2.0"
   },
   "devDependencies": {
+    "babel-preset-env": "^1.5.1",
     "babel-preset-stage-0": "^6.24.1",
     "chai": "^3.5.0",
     "chai-http": "^3.0.0",
     "faker": "^4.1.0",
-    "mocha": "^3.3.0"
+    "file-loader": "^0.11.2",
+    "mocha": "^3.3.0",
+    "redux-logger": "^3.0.6",
+    "webpack": "^2.6.1"
   }
 }
