@@ -1,7 +1,7 @@
-var React = require('react')
-var {withRouter} = require('react-router-dom')
-var {getCars} = require('./actions')
-var {connect} = require('react-redux')
+import React from 'react'
+import {withRouter} from 'react-router-dom'
+import {getCars} from './actions'
+import {connect} from 'react-redux'
 
 class CarSearch extends React.Component{
   constructor(props) {
@@ -15,8 +15,6 @@ class CarSearch extends React.Component{
     e.preventDefault()
     let query = this.refs.model.value
     let city = this.refs.city.value
-    //console.log(getCars(query, city))
-    //this.props.dispatch(getCars(query, city))
     this.props.getCars(query, city)
     if(this.props.location.pathname !== 'carlist'){
       this.props.history.push('/carlist')
@@ -28,8 +26,6 @@ class CarSearch extends React.Component{
       <div className="car-search">
         <h3>Search for your car</h3>
         <form onSubmit={this.startSearch.bind(this)}>
-          <label>State</label><br/>
-          <input type="text" placeholder="state"/>
           <label>City</label><br/>
           <input type="text" ref="city" placeholder="city"/>
           <label>Model</label><br/>
@@ -42,4 +38,5 @@ class CarSearch extends React.Component{
 }
 
 function mapStateToProps(state){return {}}
-module.exports = withRouter(connect(mapStateToProps, { getCars })(CarSearch))
+
+export default withRouter(connect(mapStateToProps, { getCars })(CarSearch))
