@@ -1,14 +1,12 @@
-var React = require('react')
+import React from 'react'
 
 class Header extends React.Component{
   render(){
     return (
-    <div>
-      <header>
-        <h2 className="hind">CarStock</h2>
-      </header>
-    </div>
+    <header>
+      <h2 className="hind">CarStock</h2>
+    </header>
   )}
 }
 
-module.exports = Header
+export default Header
