@@ -1,5 +1,5 @@
-var React = require('react')
-var reactDom = require('react-dom')
-var App = require('./App')
+import React from 'react'
+import reactDom from 'react-dom'
+import App from './App'
 
 reactDom.render(<App></App>, document.getElementById('app'))
