@@ -1,4 +1,5 @@
 import { createStore, applyMiddleware, compose } from 'redux';
+import { createLogger } from 'redux-logger'
 import thunk from 'redux-thunk';
 import rootReducer from './allReducers';
 
@@ -7,7 +8,7 @@ export default function configureStore(initialState) {
         rootReducer,
         initialState,
         compose(
-          applyMiddleware(thunk),
+          applyMiddleware(thunk, createLogger()),
           window.devToolsExtension ? window.devToolsExtension() : f => f
         ),
     );
